refactor websocket controllers and tests into each domain
